requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
transformers==4.26.1
PyPDF2==3.0.0
python-docx==0.8.11
nltk==3.8.1
requests==2.31.0
